requirements.txt
Django==4.0
django-ckeditor==6.2.0
django-cleanup==5.2.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-import-export==2.7.1
django-js-asset==1.2.2
django-markdownx==3.0.1
django-pandas==0.6.6
djangorestframework==3.13.1
channels==3.0.4
channels-redis==3.3.1
Markdown==3.3.6
matplotlib==3.5.1
mysql-connector-python==8.0.27
mysqlclient==2.1.0
numpy==1.22.0
pandas==1.3.5
python-dateutil==2.8.2
scikit-learn==1.0.2
scipy==1.7.3
stripe==2.64.0
Twisted==21.7.0
